+++ packages/ksef-client-ts/src/models/limits/types.ts
@@ -16,6 +16,7 @@ export interface EffectiveApiRateLimits {
   invoiceExport: EffectiveApiRateLimitValues;
   invoiceExportStatus: EffectiveApiRateLimitValues;
   invoiceDownload: EffectiveApiRateLimitValues;
+  collectiveIdentifier: EffectiveApiRateLimitValues;
   other: EffectiveApiRateLimitValues;
 }
 
@@ -31,9 +32,14 @@ export interface BatchSessionEffectiveContextLimits {
   maxInvoices: number;
 }
 
+export interface CollectiveIdentifierEffectiveContextLimits {
+  maxInvoices: number;
+}
+
 export interface EffectiveContextLimits {
   onlineSession: OnlineSessionEffectiveContextLimits;
   batchSession: BatchSessionEffectiveContextLimits;
+  collectiveIdentifier: CollectiveIdentifierEffectiveContextLimits;
 }
 
 export interface EnrollmentEffectiveSubjectLimits {
@@ -71,6 +77,7 @@ export interface ApiRateLimitsOverride {
   invoiceExport: ApiRateLimitValuesOverride;
   invoiceExportStatus: ApiRateLimitValuesOverride;
   invoiceDownload: ApiRateLimitValuesOverride;
+  collectiveIdentifier: ApiRateLimitValuesOverride;
   other: ApiRateLimitValuesOverride;
 }
 
@@ -86,9 +93,14 @@ export interface BatchSessionContextLimitsOverride {
   maxInvoices: number;
 }
 
+export interface CollectiveIdentifierContextLimitsOverride {
+  maxInvoices: number;
+}
+
 export interface SetSessionLimitsRequest {
   onlineSession: OnlineSessionContextLimitsOverride;
   batchSession: BatchSessionContextLimitsOverride;
+  collectiveIdentifier: CollectiveIdentifierContextLimitsOverride;
 }
 
 export interface EnrollmentSubjectLimitsOverride {